Clarify extension method declaration
Extension methods can only be declared in top-level type. They can't be declared in nested types.
