+++ docs/csharp/programming-guide/classes-and-structs/snippets/how-to-implement-and-call-a-custom-extension-method/Program.cs
@@ -1,0 +1,25 @@
+﻿using CustomExtensions;
+
+string s = "The quick brown fox jumped over the lazy dog.";
+// Call the method as if it were an
+// instance method on the type. Note that the first
+// parameter is not specified by the calling code.
+int i = s.WordCount();
+System.Console.WriteLine("Word count of s is {0}", i);
+
+
+namespace CustomExtensions
+{
+    // Extension methods must be defined in a static class.
+    public static class StringExtension
+    {
+        // This is the extension method.
+        // The first parameter takes the "this" modifier
+        // and specifies the type for which the method is defined.
+        public static int WordCount(this string str)
+        {
+            return str.Split(new char[] {' ', '.','?'}, StringSplitOptions.RemoveEmptyEntries).Length;
+        }
+    }
+}
+
